broker/requirements.txt
# Web framework
flask
gunicorn

# Docker SDK
docker

# HTTP client
requests

# PostgreSQL
psycopg2-binary

# YAML config
pyyaml
